+++ quickstart/src/index.js
@@ -7,37 +7,83 @@ var previewTracks;
 var identity;
 var roomName;
 
-// Attach the Tracks to the DOM.
+// Attach the Track to the DOM.
+function attachTrack(track, container) {
+  container.appendChild(track.attach());
+}
+
+// Attach array of Tracks to the DOM.
 function attachTracks(tracks, container) {
   tracks.forEach(function(track) {
-    container.appendChild(track.attach());
+    attachTrack(track, container);
   });
 }
 
-// Attach the Participant's Tracks to the DOM.
-function attachParticipantTracks(participant, container) {
-  var tracks = Array.from(participant.tracks.values());
-  attachTracks(tracks, container);
+// Detach given track from the DOM.
+function detachTrack(track) {
+  track.detach().forEach(function(element) {
+    element.remove();
+  });
 }
 
-// Detach the Tracks from the DOM.
-function detachTracks(tracks) {
-  tracks.forEach(function(track) {
-    track.detach().forEach(function(detachedElement) {
-      detachedElement.remove();
-    });
+// Appends remoteParticipant name to the DOM.
+function appendName(identity, container) {
+  const name = document.createElement('p');
+  name.id = `participantName-${identity}`;
+  name.className = 'instructions';
+  name.textContent = identity;
+  container.appendChild(name);
+}
+
+// Removes remoteParticipant container from the DOM.
+function removeName(participant) {
+  if (participant) {
+    let { identity } = participant;
+    const container = document.getElementById(
+      `participantContainer-${identity}`
+    );
+    container.parentNode.removeChild(container);
+  }
+}
+
+// A new RemoteTrack was published to the Room.
+function trackPublished(publication, container) {
+  if (publication.isSubscribed) {
+    attachTrack(publication.track, container);
+  }
+  publication.on('subscribed', function(track) {
+    log('Subscribed to ' + publication.kind + ' track');
+    attachTrack(track, container);
   });
+  publication.on('unsubscribed', detachTrack);
+}
+
+// A RemoteTrack was unpublished from the Room.
+function trackUnpublished(publication) {
+  log(publication.kind + ' track was unpublished.');
+}
+
+// A new RemoteParticipant joined the Room
+function participantConnected(participant, container) {
+  let selfContainer = document.createElement('div');
+  selfContainer.id = `participantContainer-${participant.identity}`;
+
+  container.appendChild(selfContainer);
+  appendName(participant.identity, selfContainer);
+
+  participant.tracks.forEach(function(publication) {
+    trackPublished(publication, selfContainer);
+  });
+  participant.on('trackPublished', function(publication) {
+    trackPublished(publication, selfContainer);
+  });
+  participant.on('trackUnpublished', trackUnpublished);
 }
 
 // Detach the Participant's Tracks from the DOM.
 function detachParticipantTracks(participant) {
-  var tracks = Array.from(participant.tracks.values());
-  detachTracks(tracks);
-}
-
-// Check for WebRTC
-if (!navigator.webkitGetUserMedia && !navigator.mozGetUserMedia) {
-  alert('WebRTC is not available in your browser.');
+  var tracks = getTracks(participant);
+  tracks.forEach(detachTrack);
 }
 
 // When we are about to transition away from this page, disconnect
@@ -81,49 +127,47 @@ $.getJSON('/token', function(data) {
   };
 });
 
+// Get the Participant's Tracks.
+function getTracks(participant) {
+  return Array.from(participant.tracks.values()).filter(function(publication) {
+      return publication.track;
+    }).map(function(publication) {
+      return publication.track;
+    });
+}
+
 // Successfully connected!
 function roomJoined(room) {
   window.room = activeRoom = room;
 
   log("Joined as '" + identity + "'");
   document.getElementById('button-join').style.display = 'none';
-  document.getElementById('button-leave').style.display = 'inline';
+  document.getElementById('button-leave').style.display = 'block';
 
   // Attach LocalParticipant's Tracks, if not already attached.
   var previewContainer = document.getElementById('local-media');
   if (!previewContainer.querySelector('video')) {
-    attachParticipantTracks(room.localParticipant, previewContainer);
+    attachTracks(getTracks(room.localParticipant), previewContainer);
   }
 
   // Attach the Tracks of the Room's Participants.
+  var remoteMediaContainer = document.getElementById('remote-media');
   room.participants.forEach(function(participant) {
     log("Already in Room: '" + participant.identity + "'");
-    var previewContainer = document.getElementById('remote-media');
-    attachParticipantTracks(participant, previewContainer);
+    participantConnected(participant, remoteMediaContainer);
   });
 
   // When a Participant joins the Room, log the event.
   room.on('participantConnected', function(participant) {
     log("Joining: '" + participant.identity + "'");
-  });
-
-  // When a Participant adds a Track, attach it to the DOM.
-  room.on('trackAdded', function(track, participant) {
-    log(participant.identity + " added track: " + track.kind);
-    var previewContainer = document.getElementById('remote-media');
-    attachTracks([track], previewContainer);
-  });
-
-  // When a Participant removes a Track, detach it from the DOM.
-  room.on('trackRemoved', function(track, participant) {
-    log(participant.identity + " removed track: " + track.kind);
-    detachTracks([track]);
+    participantConnected(participant, remoteMediaContainer);
   });
 
   // When a Participant leaves the Room, detach its Tracks.
   room.on('participantDisconnected', function(participant) {
-    log("Participant '" + participant.identity + "' left the room");
+    log("RemoteParticipant '" + participant.identity + "' left the room");
     detachParticipantTracks(participant);
+    removeName(participant);
   });
 
   // Once the LocalParticipant leaves the room, detach the Tracks
@@ -134,11 +178,13 @@ function roomJoined(room) {
       previewTracks.forEach(function(track) {
         track.stop();
       });
+      previewTracks = null;
     }
     detachParticipantTracks(room.localParticipant);
     room.participants.forEach(detachParticipantTracks);
+    room.participants.forEach(removeName);
     activeRoom = null;
-    document.getElementById('button-join').style.display = 'inline';
+    document.getElementById('button-join').style.display = 'block';
     document.getElementById('button-leave').style.display = 'none';
   });
 }
@@ -150,15 +196,16 @@ document.getElementById('button-preview').onclick = function() {
     : Video.createLocalTracks();
 
   localTracksPromise.then(function(tracks) {
-    window.previewTracks = previewTracks = tracks;
-    var previewContainer = document.getElementById('local-media');
-    if (!previewContainer.querySelector('video')) {
-      attachTracks(tracks, previewContainer);
+      window.previewTracks = previewTracks = tracks;
+      var previewContainer = document.getElementById('local-media');
+      if (!previewContainer.querySelector('video')) {
+        attachTracks(tracks, previewContainer);
+      }
+    },function(error) {
+      console.error('Unable to access local media', error);
+      log('Unable to access Camera and Microphone');
     }
-  }, function(error) {
-    console.error('Unable to access local media', error);
-    log('Unable to access Camera and Microphone');
-  });
+  );
 };
 
 // Activity log.